changing type lat and lng fields in Google maps response

+++ lib/beckn-core/src/Beckn/External/GoogleMaps/Types.hs
@@ -31,8 +31,8 @@ newtype Geometry = Geometry
   deriving (Generic, ToJSON, FromJSON)
 
 data LocationS = LocationS
-  { lat :: Int,
-    lng :: Int
+  { lat :: Double,
+    lng :: Double
   }
   deriving (Generic, ToJSON, FromJSON)
